Update patch set 1
Patch Set 1:

> > In order to validate this change, please run system/netd/tests/runtests.sh after having updated the on-device iptables copy (e.g., by flashing).
> 
> Why isn't that in presubmit, or at least in an automated system so that it can be run via forrest?

They do run on apct, but only on internal master. We're trying to get them to run in APCT on AOSP but I don't think that's (widely?) used yet.

They only run in presubmit for system/netd because we did not make the effort of adding all the project dependencies to the netd test definitions. I'm trying to do that now.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>