Update patch set 2
Patch Set 2:

Please add a 'Signed-off-by:' line.

Please prefix the commit message with ANDROID:

and in general probably call it:

ANDROID: iptables: introduce --alarm option for hard idletimer

Please provide a bit of a description about the difference between the default and hardidle timers.

(presumably something about default not counting time the device kernel is suspended, while hardidle/alarm does???)

--

Also: do you know where the code clears/memsets the memory underlying an iptables target?

ie. how do we know this field is initialized to 0 - and that it *was* initialized to zero even before this patch.

Would it be painful if we switched the --send_nl_msg option away from a 0/1 to a present/not present field?  (ie. would that cause issues)

Patch-set: 2
CC: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>