Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Please add a 'Signed-off-by:' line.
> 
> Please prefix the commit message with ANDROID:
> 
> and in general probably call it:
> 
> ANDROID: iptables: introduce --alarm option for hard idletimer
> 
> Please provide a bit of a description about the difference between the default and hardidle timers.
> 
> (presumably something about default not counting time the device kernel is suspended, while hardidle/alarm does???)
> 
> --
> 
> Also: do you know where the code clears/memsets the memory underlying an iptables target?
> 
> ie. how do we know this field is initialized to 0 - and that it *was* initialized to zero even before this patch.
> 
> Would it be painful if we switched the --send_nl_msg option away from a 0/1 to a present/not present field?  (ie. would that cause issues)

I'll update the commit text.

Regarding the question about how the newly added field would be initialized to 0, I checked the code as to how iptable options are added and executed through iptable/ip6table binaries. The iptable options are added by the client, if a particular option is not present, then it will not be added while adding the iptable rule, it doesn't work like a buffer being passed as clients can pass variable arguments. if --alarm option is not added, then this wont be added and therefore only idletimer will get added. if --alarm is added then alarm timer/hardidletimer would be added. We've tested this too.

Patch-set: 2