Update patch set 2
Patch Set 2:

Something inside the iptables binary has to either stack or heap allocate the struct that it will populate and later pass to the kernel.

We need to be certain that allocation is zeroed.

I'm pretty sure it is, because I imagine other things wouldn't work if it wasn't, but it would be good to doublecheck.

Patch-set: 2