iptables: remove a memory leak in legacy iptables
Commit 148131f20421 ("xtables: Fix for false-positive rule
matching") introduced cloning of targets to fix a bug where when
comparing two rules with non-standard targets, the comparison
would not consider the targets' payloads.

The clone is freed in nft_clear_iptables_command_state and in
ebt_cs_clean. According to the commit message, in legacy iptables
it is not possible to reuse a target, but the clone is made
anyway, and never freed.  When using a long-running legacy
iptables process (e.g., an iptables-restore) then the process
will leak memory any time two subsequent commands use the same
target.

Fix this by clearing the clones manually in legacy iptables.
This is a very simple fix but it would arguably be better to
ensure that when using legacy iptables, the clone is never made.

Bug: 162925719
Bug: 168688680
Test: manual testing with a script that continuously adds/removes rules
Fixes: 148131f20421 ("xtables: Fix for false-positive rule matching")
Change-Id: I26311833eb45c6127c28ea400e95c10b5e1f8e1e